delete unrelated code
