--- Procfile
@@ -1,2 +1,0 @@
-web: bin/blocks-metrics
-worker: bin/collector
