--- app/auth.go
@@ -1,20 +1,0 @@
-package app
-
-import (
-	"net/http"
-)
-
-//when required we'll apply auth
-var JwtAuthentication = func(next http.Handler) http.Handler {
-	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
-		notAuth := []string{"/api/blocks/{id}"}
-		//current request path
-		requestPath := r.URL.Path
-		for _, value := range notAuth {
-			if value == requestPath {
-				next.ServeHTTP(w, r)
-				return
-			}
-		}
-	})
-}
