--- heroku.yml
@@ -1,8 +1,0 @@
-# https://devcenter.heroku.com/articles/heroku-yml-build-manifest
-# Officially unsupported, but works.
-build:
-  languages:
-    - go
-
-run:
-  web: blocks-metrics
