+++ pkg/models/transaction.go
@@ -1,0 +1,6 @@
+package models
+
+type Transaction struct {
+	Hash    []byte
+	Message string
+}
